requirements.txt
actionlib==1.12.1
angles==1.9.12
base_local_planner==1.16.7
bondpy==1.8.5
camera_calibration_parsers==1.11.13
carla==0.9.13
catkin==0.7.29
catkin_virtualenv==0.6.1
clarabel==0.6.0
cmake==3.26.3
cmeel==0.49.0
contourpy==1.0.7
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.1
cvxpy==1.4.2
cvxpygen==0.3.2
cycler==0.11.0
diagnostic_updater==1.9.7
dynamic_reconfigure==1.6.5
ecos==2.0.12
fonttools==4.39.3
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
geodesy==0.5.3
gps_common==0.3.4
image_geometry==1.13.1
interactive-markers==1.11.5
jsk-recognition-utils==1.2.15
jsk_rviz_plugins==2.1.8
jsk_topic_tools==2.2.12
kiwisolver==1.4.4
laser_geometry==1.6.7
matplotlib==3.7.1
message_filters==1.14.13
nmea_navsat_driver==0.5.2
numpy==1.24.3
osqp==0.6.2.post9
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
proxsuite==0.4.1
pybind11==2.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python_qt_binding==0.4.4
pytz==2023.3
PyYAML==6.0.1
qdldl==0.1.7
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-app==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
rosapi==0.11.16
rosbag==1.14.13
rosboost-cfg==1.14.9
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rospy_message_converter==0.5.9
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-image-view==0.4.17
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_tf_tree==0.6.0
rviz==1.13.30
scipy==1.10.1
scs==3.2.3
sensor-msgs==1.12.8
shapely==2.0.3
six==1.16.0
smclib==1.8.5
sound-play==0.3.17
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf2_sensor_msgs==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.13
turtlebot3_autorace_core==1.2.0
tzdata==2023.3
unique_id==1.0.6
urdfdom-py==0.4.6
xacro==1.13.20
